requirements.txt
aiohttp
beautifulsoup4
hachoir
numpy
Pillow
pyrogram
requests
tgcrypto
youtube-dl
olefile